Add a test dir, test bin/fullpath
Hopefully having this in place will let me begin doing this sort of
thing in the future, b/c depending on the script, it may be easiest to
test it from the outside. (others, it makes more sense to test inside,
and then, of course, you can always do both).